Fixed compiler warning [skip ci].

+++ src/libxsmm_gemm.c
@@ -1571,7 +1571,7 @@ LIBXSMM_API void libxsmm_gemm_internal_set_batchflag(libxsmm_gemm_descriptor* de
         const libxsmm_blasint typesize = LIBXSMM_TYPESIZE(oprec);
         const libxsmm_blasint csize = descriptor->ldc * descriptor->n * typesize;
         /* finalize assumption if matrix-size is a multiple of the vector-width */
-        descriptor->flags |= (0 == LIBXSMM_MOD2(csize, vw) ? LIBXSMM_GEMM_FLAG_ALIGN_C_NTS_HINT : 0);
+        descriptor->flags |= (unsigned short)(0 == LIBXSMM_MOD2(csize, vw) ? LIBXSMM_GEMM_FLAG_ALIGN_C_NTS_HINT : 0);
       }
     }
 #if defined(LIBXSMM_GEMM_BATCHREDUCE)
